+++ Coimbra.Editor/TypeDropdownItem.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using UnityEditor.IMGUI.Controls;
+
+namespace Coimbra.Editor
+{
+    internal sealed class TypeDropdownItem : AdvancedDropdownItem
+    {
+        internal readonly Type Type;
+
+        public TypeDropdownItem(Type type, string name)
+            : base(name)
+        {
+            Type = type;
+        }
+    }
+}