+++ Coimbra/TypeDropdownAttribute.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace Coimbra
+{
+    /// <summary>
+    /// Apply this to a field with <see cref="SerializeReference"/> to expose a type dropdown for it.
+    /// </summary>
+    /// 
+    public sealed class TypeDropdownAttribute : PropertyAttribute { }
+}